Many updates since the last source commit... - Updated parameter system to allow use of MatrixArrayParameters. - Added generalized animation classes. - Added SkinnedActor class to demonstrate use of vertex skinning in the engine. - Added additional MS3D model loader for animation data. - Added debugging visualizations for scene nodes. - Cleaned up many TODO:'s... - Added Texture2D visualization actor for viewing resources at runtime in the 3D scene. - Updated Actor class to hold an array of Entity3D objects that get added to it. - Added ability to have generalized render parameters to both entities and materials, allowing for very flexible rendering scenario construction (tested with the SkinnedActor). - Added some additional GeometryGenerator functions.

+++ trunk/Hieroglyph3/Applications/BasicRenderViews/App.cpp
@@ -162,8 +162,8 @@ void App::Initialize()
 	m_pRenderView = new ViewPerspective( *m_pRenderer11, m_RenderTarget, m_DepthTarget );
 	m_pRenderView->SetBackColor( Vector4f( 0.6f, 0.6f, 0.6f, 0.6f ) );
 
-	D3DXMatrixPerspectiveFovLH( (D3DXMATRIX*)&m_pRenderView->ProjMatrix, D3DX_PI/4, 
-		640.0f / 320.0f, 0.1f, 100.0f );
+	D3DXMatrixPerspectiveFovLH( (D3DXMATRIX*)&m_pRenderView->ProjMatrix, 
+		static_cast<float>( D3DX_PI ) / 4, 640.0f / 320.0f, 0.1f, 100.0f );
 
 	m_pRoot = new Node3D();
 
@@ -177,7 +177,7 @@ void App::Initialize()
 		m_pEntity[i] = new Entity3D();
 		m_pEntity[i]->SetGeometry( m_pGeometry );
 		m_pEntity[i]->SetMaterial( m_pMaterial, false );
-		m_pEntity[i]->Position() = Vector3f( i * 4, 4.0f * ( i % 2 ) - 2.0f, 0.0f );
+		m_pEntity[i]->Position() = Vector3f( static_cast<float>( i ) * 4, 4.0f * ( i % 2 ) - 2.0f, 0.0f );
 
 		m_pRoot->AttachChild( m_pEntity[i] );
 	}
